+++ src/rules/types.ts
@@ -11,7 +11,7 @@ const get = (mergedOptions: MergedOptions): Linter.FlatConfig[] => {
         mergedOptions.base?.types && {
             name: "Base pure typescript",
             plugins: {
-                // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment,@typescript-eslint/no-explicit-any
+                // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment,@typescript-eslint/no-explicit-any,@typescript-eslint/no-unsafe-type-assertion
                 "@typescript-eslint": typescript as any,
             },
             rules: {